Addd Item

+++ Assets/!TeamMenber/_marsh/Scripts/Item/WeaponBase.cs
@@ -6,10 +6,10 @@ public class WeaponBase : ItemBase
 {
     public int weaponAttack;
     public int weaponCriticalChance;
-    public int weaponCriticalDamage;
+    public float weaponCriticalDamage;
 
     // コンストラクタ：新しいアイテムを作成
-    public WeaponBase(string name, int id,int atk,int criChance, int criDamage, Sprite image = null, int skill1 = 0,int skill2 = 0)
+    public WeaponBase(string name, int id,int atk,int criChance, float criDamage, Sprite image = null, int skill1 = 0,int skill2 = 0)
         : base(name, id, eItemType.Weapon, image, skill1, skill2) {
 
         weaponAttack = atk;
